fix(KUI-1140): structural changes
utils-shared folders have been added, some refactoring
