+++ public/js/app/util/helpers.js
@@ -48,18 +48,4 @@ const formatRoundName = (language, shortName, semester, applicationCode) => {
       }-${applicationCode}`
 }
 
-function convertLangToIndex(langShortName) {
-  return langShortName === 'en' ? 0 : 1
-}
-
-const seasonStr = (language, semesterRaw) => {
-  if (!semesterRaw) return ''
-  const isLangANumber = typeof language === 'number'
-  const langIndex = isLangANumber ? language : convertLangToIndex(language)
-  const { course_short_semester } = i18n.messages[langIndex].messages
-  const termStringAsSeason = `${course_short_semester[semesterRaw.toString()[4]]}${semesterRaw.toString().slice(0, 4)}`
-
-  return termStringAsSeason
-}
-
-export { formatDate, getTodayDate, getDateFormat, isValidDate, formatRoundName, seasonStr, convertLangToIndex }
+export { formatDate, getTodayDate, getDateFormat, isValidDate, formatRoundName }
