+++ server/utils-shared/README.md
@@ -1,0 +1,5 @@
+This folder contains candidates for a shared utils lib in server folder. 
+
+Beware that utils here might be duplicated in other Om kursen repos.
+
+Read more about this (temporary?) way to share code in confluence: https://confluence.sys.kth.se/confluence/x/6wYJDQ
